add vue router factor catcolor in separate file add routing for permalink + permalink handling
